# packages/frontend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "德州扑克游戏API",
    "version": "1.0.0",
    "description": "德州扑克游戏后端API文档"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "本地开发服务器"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "GameRoom": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "房间ID"
          },
          "name": {
            "type": "string",
            "description": "房间名称"
          },
          "owner_id": {
            "type": "string",
            "description": "房主ID"
          },
          "room_type": {
            "type": "string",
            "enum": [
              "public",
              "private"
            ],
            "description": "房间类型"
          },
          "small_blind": {
            "type": "number",
            "description": "小盲注"
          },
          "big_blind": {
            "type": "number",
            "description": "大盲注"
          },
          "max_players": {
            "type": "number",
            "description": "最大玩家数"
          },
          "current_players": {
            "type": "number",
            "description": "当前玩家数"
          },
          "game_status": {
            "type": "string",
            "enum": [
              "waiting",
              "playing",
              "finished"
            ],
            "description": "游戏状态"
          },
          "min_buy_in": {
            "type": "number",
            "description": "最小买入"
          },
          "max_buy_in": {
            "type": "number",
            "description": "最大买入"
          },
          "table_type": {
            "type": "string",
            "enum": [
              "cash",
              "tournament"
            ],
            "description": "桌子类型"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "创建时间"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "更新时间"
          }
        },
        "required": [
          "id",
          "name",
          "owner_id",
          "room_type",
          "small_blind",
          "big_blind",
          "max_players",
          "current_players",
          "game_status",
          "table_type",
          "created_at",
          "updated_at"
        ]
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/rooms": {
      "post": {
        "summary": "创建房间",
        "tags": [
          "Rooms"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "maxPlayers": {
                    "type": "number"
                  },
                  "buyIn": {
                    "type": "number"
                  },
                  "smallBlind": {
                    "type": "number"
                  },
                  "bigBlind": {
                    "type": "number"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "房间创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "room": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        },
                        "hostId": {
                          "type": "string"
                        },
                        "players": {
                          "type": "array",
                          "items": {
                            "type": "object"
                          }
                        },
                        "maxPlayers": {
                          "type": "number"
                        },
                        "buyIn": {
                          "type": "number"
                        },
                        "smallBlind": {
                          "type": "number"
                        },
                        "bigBlind": {
                          "type": "number"
                        },
                        "status": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "创建房间失败"
          },
          "401": {
            "description": "未授权"
          }
        }
      },
      "get": {
        "summary": "获取房间列表",
        "tags": [
          "Rooms"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string"
            },
            "description": "房间状态筛选"
          }
        ],
        "responses": {
          "200": {
            "description": "获取房间列表成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "rooms": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "hostId": {
                            "type": "string"
                          },
                          "playerCount": {
                            "type": "number"
                          },
                          "maxPlayers": {
                            "type": "number"
                          },
                          "buyIn": {
                            "type": "number"
                          },
                          "smallBlind": {
                            "type": "number"
                          },
                          "bigBlind": {
                            "type": "number"
                          },
                          "status": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "获取房间列表失败"
          }
        }
      }
    },
    "/rooms/validate-password/{roomId}": {
      "post": {
        "summary": "验证私人房间密码",
        "tags": [
          "Rooms"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "roomId",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "房间ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "password": {
                    "type": "string",
                    "description": "房间密码"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "密码验证成功"
          },
          "401": {
            "description": "密码错误或未授权"
          },
          "404": {
            "description": "房间不存在"
          }
        }
      }
    },
    "/rooms/{roomId}": {
      "put": {
        "summary": "更新房间信息",
        "tags": [
          "Rooms"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "roomId",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "房间ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "small_blind": {
                    "type": "number"
                  },
                  "big_blind": {
                    "type": "number"
                  },
                  "max_players": {
                    "type": "number"
                  },
                  "password": {
                    "type": "string"
                  },
                  "min_buy_in": {
                    "type": "number"
                  },
                  "max_buy_in": {
                    "type": "number"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "房间更新成功"
          },
          "401": {
            "description": "未授权或不是房间所有者"
          },
          "404": {
            "description": "房间不存在"
          }
        }
      },
      "delete": {
        "summary": "删除房间",
        "tags": [
          "Rooms"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "roomId",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "房间ID"
          }
        ],
        "responses": {
          "200": {
            "description": "房间删除成功"
          },
          "401": {
            "description": "未授权或不是房间所有者"
          },
          "404": {
            "description": "房间不存在"
          }
        }
      },
      "get": {
        "summary": "获取房间详情",
        "tags": [
          "Rooms"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "roomId",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "房间ID"
          }
        ],
        "responses": {
          "200": {
            "description": "获取房间详情成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "room": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        },
                        "hostId": {
                          "type": "string"
                        },
                        "players": {
                          "type": "array",
                          "items": {
                            "type": "object"
                          }
                        },
                        "maxPlayers": {
                          "type": "number"
                        },
                        "buyIn": {
                          "type": "number"
                        },
                        "smallBlind": {
                          "type": "number"
                        },
                        "bigBlind": {
                          "type": "number"
                        },
                        "status": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "获取房间详情失败"
          },
          "401": {
            "description": "未授权"
          },
          "404": {
            "description": "房间不存在"
          }
        }
      }
    },
    "/rooms/user/created": {
      "get": {
        "summary": "获取用户创建的房间列表",
        "tags": [
          "Rooms"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "获取用户创建的房间列表成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "rooms": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/GameRoom"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "获取房间列表失败"
          }
        }
      }
    },
    "/rooms/{roomId}/join": {
      "post": {
        "summary": "加入房间",
        "tags": [
          "Rooms"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "roomId",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "房间ID"
          }
        ],
        "responses": {
          "200": {
            "description": "加入房间成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "加入房间失败"
          },
          "401": {
            "description": "未授权或密码错误"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "password": {
                    "type": "string",
                    "description": "私人房间密码"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/rooms/{roomId}/leave": {
      "post": {
        "summary": "离开房间",
        "tags": [
          "Rooms"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "roomId",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "房间ID"
          }
        ],
        "responses": {
          "200": {
            "description": "离开房间成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "离开房间失败"
          },
          "401": {
            "description": "未授权"
          }
        }
      }
    },
...
  },
...
}